defining anonymous types
