+++ src/core/generator/openapi/mod.rs
@@ -3,5 +3,6 @@ pub mod helpers;
 mod query_generator;
 mod type_generator;
 
+pub use anonymous_type_generator::{AnonymousTypeGenerator, AnonymousTypes};
 pub use query_generator::QueryGenerator;
 pub use type_generator::TypeGenerator;
